+++ test/test_revparse.py
@@ -35,47 +35,47 @@ PARENT_SHA = '5ebeeebb320790caf276b9fc8b24546d63316533'  # HEAD^
 
 def test_revparse_single(testrepo):
     o = testrepo.revparse_single('HEAD')
-    assert o.hex == HEAD_SHA
+    assert str(o.id) == HEAD_SHA
 
     o = testrepo.revparse_single('HEAD^')
-    assert o.hex == PARENT_SHA
+    assert str(o.id) == PARENT_SHA
 
     o = testrepo.revparse_single('@{-1}')
-    assert o.hex == '5470a671a80ac3789f1a6a8cefbcf43ce7af0563'
+    assert str(o.id) == '5470a671a80ac3789f1a6a8cefbcf43ce7af0563'
 
 
 def test_revparse_ext(testrepo):
     o, r = testrepo.revparse_ext('master')
-    assert o.hex == HEAD_SHA
+    assert str(o.id) == HEAD_SHA
     assert r == testrepo.references['refs/heads/master']
 
     o, r = testrepo.revparse_ext('HEAD^')
-    assert o.hex == PARENT_SHA
+    assert str(o.id) == PARENT_SHA
     assert r is None
 
     o, r = testrepo.revparse_ext('i18n')
-    assert o.hex.startswith('5470a67')
+    assert str(o.id).startswith('5470a67')
     assert r == testrepo.references['refs/heads/i18n']
 
 
 def test_revparse_1(testrepo):
     s = testrepo.revparse('master')
-    assert s.from_object.hex == HEAD_SHA
+    assert str(s.from_object.id) == HEAD_SHA
     assert s.to_object is None
     assert s.flags == RevSpecFlag.SINGLE
 
 
 def test_revparse_range_1(testrepo):
     s = testrepo.revparse('HEAD^1..acecd5e')
-    assert s.from_object.hex == PARENT_SHA
-    assert s.to_object.hex.startswith('acecd5e')
+    assert str(s.from_object.id) == PARENT_SHA
+    assert str(s.to_object.id).startswith('acecd5e')
     assert s.flags == RevSpecFlag.RANGE
 
 
 def test_revparse_range_2(testrepo):
     s = testrepo.revparse('HEAD...i18n')
-    assert s.from_object.hex.startswith('2be5719')
-    assert s.to_object.hex.startswith('5470a67')
+    assert str(s.from_object.id).startswith('2be5719')
+    assert str(s.to_object.id).startswith('5470a67')
     assert s.flags == RevSpecFlag.RANGE | RevSpecFlag.MERGE_BASE
     assert testrepo.merge_base(s.from_object.id, s.to_object.id) is not None